Activate Touringcar Messages (#100)
As (dummy) Touringcar Messages only exist on Acceptance and Production,
and they cannot be accessed locally due to CORS, this PR needs to be
merged with `azure` to get the code on Acceptance.

You might notice the `ts-ignore` comments everywhere. They are added due
to an upstream type error. See [this bug report of
mine](https://github.com/Amsterdam/amsterdam-react-maps/issues/2184) for
more details.

---------

Co-authored-by: Kris de Groot <kris.degroot@teamrockstars.nl>
Co-authored-by: jjbeekman <jonathanjeroen@gmail.com>
Co-authored-by: Jeroen Beekman <j.beekman@amsterdam.nl>
Co-authored-by: jjbeekman <70761835+jjbeekman@users.noreply.github.com>
Co-authored-by: dependabot-vakgroep-so[bot] <144998647+dependabot-vakgroep-so[bot]@users.noreply.github.com>
Co-authored-by: github-merge-queue <github-merge-queue@users.noreply.github.com>
Co-authored-by: thomasbrockmeier-ams <86249927+thomasbrockmeier-ams@users.noreply.github.com>
Co-authored-by: jjbeekman <jjbeekman@users.noreply.github.com>